Events moved to trait
