--- src/main/java/de/htwg/se/connectfour/mvc/view/Event.scala
@@ -1,15 +1,0 @@
-package de.htwg.se.connectfour.mvc.view
-
-import scala.swing.event.Event
-
-class PlayerGridChanged extends Event
-
-class GridChanged extends Event
-
-class PlayerWon extends Event
-
-class Draw extends Event
-
-class FilledColumn extends Event
-
-class InvalidMove extends Event
